initial clang-tidy linting
