+++ .clang-tidy
@@ -1,0 +1,3 @@
+---
+Checks: '-*'
+WarningsAsErrors: '*'
