+++ BUILD.bazel
@@ -1,9 +1,15 @@
+load("@hedron_compile_commands//:refresh_compile_commands.bzl", "refresh_compile_commands")
+
 package(default_visibility = ["//visibility:public"])
 
 exports_files([
     "AUTHORS",
     "LICENSE",
 ])
+
+refresh_compile_commands(
+    name = "refresh_compile_commands",
+)
 
 config_setting(
     name = "msvc",
