+++ MODULE.bazel
@@ -25,7 +25,7 @@ bazel_dep(name = "hedron_compile_commands", dev_dependency = True)
 git_override(
     module_name = "hedron_compile_commands",
     remote = "https://github.com/hedronvision/bazel-compile-commands-extractor.git",
-    commit = "a14ad3a64e7bf398ab48105aaa0348e032ac87f8",
+    commit = "4f28899228fb3ad0126897876f147ca15026151e",
 )
 
 # Windows - allows building with the EWDK, must set EWDKDIR environment variable
