+++ radiant/Iterator.h
@@ -14,8 +14,10 @@
 
 #pragma once
 
-#include "radiant/TotallyRad.h"
+#include "radiant/TypeTraits.h"
 #include "radiant/Utility.h"
+
+#include <stddef.h>
 
 namespace rad
 {
