+++ radiant/Res.h
@@ -17,6 +17,8 @@
 #include "radiant/TotallyRad.h"
 #include "radiant/Result.h"
 
+#include <stdint.h>
+
 namespace rad
 {
