+++ radiant/ScopeExit.h
@@ -15,7 +15,8 @@
 #pragma once
 
 #include "radiant/TotallyRad.h"
-#include "radiant/Utility.h"
+#include "radiant/TypeTraits.h"
+#include "radiant/Utility.h" // NOLINT(misc-include-cleaner)
 
 namespace rad
 {
