+++ radiant/TypeWrapper.h
@@ -15,6 +15,7 @@
 #pragma once
 
 #include "radiant/TotallyRad.h"
+#include "radiant/TypeTraits.h"
 #include "radiant/Utility.h"
 
 #if RAD_ENABLE_STD
