+++ test/test_EmptyOptimizedPair.cpp
@@ -12,12 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "gtest/gtest.h"
-
-#include "test/TestAlloc.h"
-
 #include "radiant/EmptyOptimizedPair.h"
 
+#include "gtest/gtest.h"
+#include "test/TestAlloc.h"
 #include <vector>
 
 struct Empty
