+++ test/test_Integer.cpp
@@ -12,8 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "gtest/gtest.h"
 #include "radiant/Integer.h"
+
+#include "gtest/gtest.h"
 
 // clang-format off
