+++ test/test_ScopeExit.cpp
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "gtest/gtest.h"
-
 #include "radiant/ScopeExit.h"
+
+#include "gtest/gtest.h"
 
 namespace
 {
