+++ test/test_StdTypeTraits.cpp
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "gtest/gtest.h"
-
-#include "radiant/detail/StdTypeTraits.h"
 #include "radiant/detail/Meta.h"
+#include "radiant/detail/StdTypeTraits.h"
+
+#include "gtest/gtest.h"
 
 namespace rad
 {
