+++ tools/rad/pyproject.toml
@@ -7,6 +7,7 @@ authors = [{ name = "Radiant Authors" }]
 dependencies = [
     "pre-commit",
     "clang-format",
+    "clang-tidy",
     "pylint",
     "black",
     "lcov_cobertura",
